Removes unused fixtures; Moves fixtures to the test modules where they are used
