+++ klax/nn/__init__.py
@@ -14,6 +14,8 @@
 
 from ._icnn import FICNN as FICNN
 from ._icnn import PICNN as PICNN
+from ._icnn import FICNNLayer as FICNNLayer
+from ._icnn import PICNNLayer as PICNNLayer
 from ._linear import InputSplitLinear as InputSplitLinear
 from ._linear import Linear as Linear
 from ._matrices import ConstantMatrix as ConstantMatrix
