+++ tests/conftest.py
@@ -30,125 +30,9 @@ jax.config.update("jax_numpy_rank_promotion", "raise")
 
 
 @pytest.fixture
-def allow_rank_promotion():
-    """Disable the jax_numpy_rank_promotion for the current test."""
-    old = jax.config.jax_numpy_rank_promotion
-    jax.config.update("jax_numpy_rank_promotion", "allow")
-    try:
-        yield
-    finally:
-        jax.config.update("jax_numpy_rank_promotion", old)
-
-
-@pytest.fixture
 def getkey():
     # Delayed import so that jaxtyping can transform the AST of Equinox before
     # it is imported, but conftest.py is ran before then.
     import equinox.internal as eqxi
 
     return eqxi.GetKey()
-
-
-@pytest.fixture
-def getzerowrap():
-    import klax
-
-    class ZeroWrapper(klax.Unwrappable[Array]):
-        """A dummy wrapper that sets all parameters to zero."""
-
-        parameter: Array
-
-        def unwrap(self) -> Array:
-            return jnp.zeros_like(self.parameter)
-
-    return ZeroWrapper
-
-
-@pytest.fixture
-def getarraywrap():
-    from typing import Self
-
-    import equinox as eqx
-
-    import klax
-
-    class Wrapper(klax.Constraint):
-        """A constraint that multiplies the parameter by 2 when applied."""
-
-        parameter: Array
-
-        def unwrap(self) -> Array:
-            return self.parameter
-
-        def apply(self) -> Self:
-            return eqx.tree_at(
-                lambda x: x.parameter,
-                self,
-                replace_fn=lambda x: 2 * x,
-            )
-
-    return Wrapper
-
-
-@pytest.fixture
-def dummy_model():
-    import equinox as eqx
-
-    class Model(eqx.Module):
-        def __call__(self, x):
-            return x
-
-    return Model()
-
-
-@pytest.fixture
-def dummy_data():
-    import jax.random as jr
-
-    x = jr.uniform(jr.key(0), (100, 4))
-    y = jnp.abs(x).sum(axis=1, keepdims=True)
-    return x, y
-
-
-@pytest.fixture
-def dummy_train_state(dummy_model):
-    import equinox as eqx
-    import optax
-
-    from klax import TrainingState
-
-    opt_state = optax.adam(1e-3).init(
-        eqx.filter(dummy_model, eqx.is_inexact_array)
-    )
-    return TrainingState.create(model=dummy_model, opt_state=opt_state)
-
-
-@pytest.fixture
-def dummy_train_static(getkey, dummy_data):
-    import optax
-
-    from klax import TrainingStatic, batch_data, mse
-
-    return TrainingStatic(
-        optimizer=optax.adam(1e-3),
-        batcher=batch_data(
-            dummy_data, batch_size=10, batch_axes=0, key=getkey()
-        ),
-        batch_axes=0,
-        loss=mse,
-        steps=1000,
-    )
-
-
-@pytest.fixture
-def dummy_metric_fn():
-    import equinox as eqx
-
-    def parameter_sum_metric(model):
-        return sum(
-            x.sum()
-            for x in jax.tree_util.tree_leaves(model)
-            if eqx.is_array(x)
-        )
-
-    return parameter_sum_metric
