+++ tests/test_wrappers.py
@@ -19,6 +19,7 @@ import jax
 import jax.numpy as jnp
 import jax.random as jr
 import pytest
+from jaxtyping import Array
 
 from klax import (
     NonNegative,
@@ -33,6 +34,32 @@ from klax import (
     non_trainable,
     unwrap,
 )
+
+
+@pytest.fixture
+def getarraywrap():
+    from typing import Self
+
+    import equinox as eqx
+
+    import klax
+
+    class Wrapper(klax.Constraint):
+        """A constraint that multiplies the parameter by 2 when applied."""
+
+        parameter: Array
+
+        def unwrap(self) -> Array:
+            return self.parameter
+
+        def apply(self) -> Self:
+            return eqx.tree_at(
+                lambda x: x.parameter,
+                self,
+                replace_fn=lambda x: 2 * x,
+            )
+
+    return Wrapper
 
 
 def test_unwrap():
